+++ src/test/java/org/mercycorps/translationcards/txcmaker/model/DeckTest.java
@@ -1,6 +1,7 @@
 package org.mercycorps.translationcards.txcmaker.model;
 
 import org.junit.Test;
+import org.mercycorps.translationcards.txcmaker.model.deck.Deck;
 
 import java.util.ArrayList;
 
@@ -10,7 +11,7 @@ import static org.junit.Assert.assertThat;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
-public class NewDeckTest {
+public class DeckTest {
 
     private final ArrayList<Translation> translations = new ArrayList<>();
     private final ArrayList<String> destinationLanguageNames = new ArrayList<>();
@@ -18,7 +19,7 @@ public class NewDeckTest {
 
     @Test
     public void shouldBeValidWithNoTranslations() {
-        NewDeck deck = new NewDeck(null, null, null, 0L, false, null, null, null, parsingErrors, translations, destinationLanguageNames);
+        Deck deck = new Deck(null, null, null, 0L, false, null, null, null, parsingErrors, translations, destinationLanguageNames);
 
         assertThat(deck.isValid(), is(true));
     }
@@ -28,7 +29,7 @@ public class NewDeckTest {
         Translation validTranslation = mock(Translation.class);
         when(validTranslation.isValid()).thenReturn(true);
         translations.add(validTranslation);
-        NewDeck deck = new NewDeck(null, null, null, 0L, false, null, null, null, parsingErrors, translations, destinationLanguageNames);
+        Deck deck = new Deck(null, null, null, 0L, false, null, null, null, parsingErrors, translations, destinationLanguageNames);
 
         assertThat(deck.isValid(), is(true));
     }
@@ -38,7 +39,7 @@ public class NewDeckTest {
         Translation validTranslation = mock(Translation.class);
         when(validTranslation.isValid()).thenReturn(false);
         translations.add(validTranslation);
-        NewDeck deck = new NewDeck(null, null, null, 0L, false, null, null, null, parsingErrors, translations, destinationLanguageNames);
+        Deck deck = new Deck(null, null, null, 0L, false, null, null, null, parsingErrors, translations, destinationLanguageNames);
 
         assertThat(deck.isValid(), is(false));
     }
@@ -49,7 +50,7 @@ public class NewDeckTest {
         when(translation.isValid()).thenReturn(true).thenReturn(false);
         translations.add(translation);
         translations.add(translation);
-        NewDeck deck = new NewDeck(null, null, null, 0L, false, null, null, null, parsingErrors, translations, destinationLanguageNames);
+        Deck deck = new Deck(null, null, null, 0L, false, null, null, null, parsingErrors, translations, destinationLanguageNames);
 
         assertThat(deck.isValid(), is(false));
     }
@@ -65,7 +66,7 @@ public class NewDeckTest {
         translations.add(helloTranslation);
         translations.add(goodbyeTranslation);
 
-        NewDeck deck = new NewDeck(null, null, null, 0L, false, null, null, null, parsingErrors, translations, destinationLanguageNames);
+        Deck deck = new Deck(null, null, null, 0L, false, null, null, null, parsingErrors, translations, destinationLanguageNames);
 
         assertThat(deck.getTranslationForSourcePhrase("Hello"), is(helloTranslation));
     }
